Add `init_with_sender` method for HTTP and gRPC
This allows creating a gRPC / HTTP server node passing an existing channel to a
node that will receive invocations.

I expect that as we move towards a model in which the main node sets up
other nodes, but does not directly process requests, this new method
will be preferred instead of the existing one that also creates a local
channel from which to read invocations, since the main node would not
have any use for it.

Ref #1639